eframe

Module epi

Platform-agnostic interface for writing apps using [egui] (epi = egui programming interface).

epi provides interfaces for window management and serialization.

Start by looking at the App trait, and implement App::update.

Structs

CreationContext

Data that is passed to AppCreator that can be used to setup and initialize your app.

Frame

Represents the surroundings of your app.

IntegrationInfo

Information about the integration passed to the use app each frame.

NativeOptions

Options controlling the behavior of a native window.

Enums

HardwareAcceleration

Selects the level of hardware graphics acceleration.

WebGlContextOption

WebGL Context options

Constants

APP_KEY

Storage key used for app

Traits

App

Implement this trait to write apps that can be compiled for both web/wasm and desktop/native using eframe.

Storage

A place where you can store custom data in a way that persists when you restart the app.

Type Aliases

AppCreator

This is how your app is created.

DynError 🔒