eframe::epi

Struct NativeOptions

pub struct NativeOptions {
    pub viewport: ViewportBuilder,
    pub vsync: bool,
    pub multisampling: u16,
    pub depth_buffer: u8,
    pub stencil_buffer: u8,
    pub hardware_acceleration: HardwareAcceleration,
    pub run_and_return: bool,
    pub centered: bool,
    pub persist_window: bool,
    pub persistence_path: Option<PathBuf>,
    pub dithering: bool,
}

Options controlling the behavior of a native window.

Additional windows can be opened using (egui viewports)[egui::viewport].

Set the window title and size using Self::viewport.

Application id

[egui::ViewportBuilder::with_app_id] is used for determining the folder to persist the app to.

On native the path is picked using [crate::storage_dir].

If you don’t set an app id, the title argument to [crate::run_native] will be used as app id instead.

Fields
viewport: ViewportBuilder

Controls the native window of the root viewport.

This is where you set things like window title and size.

If you don’t set an icon, a default egui icon will be used. To avoid this, set the icon to [egui::IconData::default].

vsync: bool

Turn on vertical syncing, limiting the FPS to the display refresh rate.

The default is true.

multisampling: u16

Set the level of the multisampling anti-aliasing (MSAA).

Must be a power-of-two. Higher = more smooth 3D.

A value of 0 turns it off (default).

egui already performs anti-aliasing via “feathering” (controlled by [egui::epaint::TessellationOptions]), but if you are embedding 3D in egui you may want to turn on multisampling.

depth_buffer: u8

Sets the number of bits in the depth buffer.

egui doesn’t need the depth buffer, so the default value is 0.

stencil_buffer: u8

Sets the number of bits in the stencil buffer.

egui doesn’t need the stencil buffer, so the default value is 0.

hardware_acceleration: HardwareAcceleration

Specify whether or not hardware acceleration is preferred, required, or not.

Default: HardwareAcceleration::Preferred.

run_and_return: bool

This controls what happens when you close the main eframe window.

If true, execution will continue after the eframe window is closed. If false, the app will close once the eframe window is closed.

This is true by default, and the false option is only there so we can revert if we find any bugs.

This feature was introduced in https://github.com/emilk/egui/pull/1889.

When true, [winit::platform::run_on_demand::EventLoopExtRunOnDemand] is used. When false, [winit::event_loop::EventLoop::run] is used.

centered: bool

On desktop: make the window position to be centered at initialization.

Platform specific:

Wayland desktop currently not supported.

persist_window: bool

Controls whether or not the native window position and size will be persisted (only if the “persistence” feature is enabled).

persistence_path: Option<PathBuf>

The folder where eframe will store the app state. If not set, eframe will use a default data storage path for each target system.

dithering: bool

Controls whether to apply dithering to minimize banding artifacts.

Dithering assumes an sRGB output and thus will apply noise to any input value that lies between two 8bit values after applying the sRGB OETF function, i.e. if it’s not a whole 8bit value in “gamma space”. This means that only inputs from texture interpolation and vertex colors should be affected in practice.

Defaults to true.