eframe

Struct CreationContext

pub struct CreationContext<'s> {
    pub egui_ctx: Context,
    pub integration_info: IntegrationInfo,
    pub storage: Option<&'s dyn Storage>,
    pub gl: Option<Arc<Context>>,
    pub get_proc_address: Option<&'s dyn Fn(&CStr) -> *const c_void>,
    /* private fields */
}

Data that is passed to AppCreator that can be used to setup and initialize your app.

Fields
egui_ctx: Context

The egui Context.

You can use this to customize the look of egui, e.g to call egui::Context::set_fonts, egui::Context::set_visuals_of etc.

integration_info: IntegrationInfo

Information about the surrounding environment.

storage: Option<&'s dyn Storage>

You can use the storage to restore app state(requires the โpersistenceโ feature).

gl: Option<Arc<Context>>

The glow::Context allows you to initialize OpenGL resources (e.g. shaders) that you might want to use later from a egui::PaintCallback.

Only available when compiling with the glow feature and using Renderer::Glow.

get_proc_address: Option<&'s dyn Fn(&CStr) -> *const c_void>

The get_proc_address wrapper of underlying GL context