Struct eframe::Frame

pub struct Frame { /* private fields */ }

Represents the surroundings of your app.

It provides methods to inspect the surroundings (are we on the web?), allocate textures, and change settings (e.g. window size).

Implementations

impl Frame

pub fn is_web(&self) -> bool

True if you are in a web environment.

Equivalent to cfg!(target_arch = "wasm32")

pub fn info(&self) -> &IntegrationInfo

Information about the integration.

pub fn storage(&self) -> Option<&dyn Storage>

A place where you can store custom data in a way that persists when you restart the app.

pub fn request_screenshot(&mut self)

Request the current frame’s pixel data. Needs to be retrieved by calling Frame::screenshot during App::post_rendering.

pub fn cancel_screenshot_request(&mut self)

Cancel a request made with Frame::request_screenshot.

pub fn screenshot(&self) -> Option<ColorImage>

During App::post_rendering, use this to retrieve the pixel data that was requested during App::update via Frame::request_screenshot.

Returns None if:

Called in App::update

Frame::request_screenshot wasn’t called on this frame during App::update

The rendering backend doesn’t support this feature (yet). Currently implemented for wgpu and glow, but not with wasm as target.

Wgpu’s GL target is active (not yet supported)

Retrieving the data was unsuccessful in some way.

See also egui::ColorImage::region

Example generating a capture of everything within a square of 100 pixels located at the top left of the app and saving it with the image crate:

struct MyApp;

impl eframe::App for MyApp {
    fn update(&mut self, ctx: &egui::Context, frame: &mut eframe::Frame) {
        // In real code the app would render something here
        frame.request_screenshot();
        // Things that are added to the frame after the call to
        // request_screenshot() will still be included.
    }

    fn post_rendering(&mut self, _window_size: [u32; 2], frame: &eframe::Frame) {
        if let Some(screenshot) = frame.screenshot() {
            let pixels_per_point = frame.info().native_pixels_per_point;
            let region = egui::Rect::from_two_pos(
                egui::Pos2::ZERO,
                egui::Pos2{ x: 100., y: 100. },
            );
            let top_left_corner = screenshot.region(&region, pixels_per_point);
            image::save_buffer(
                "top_left.png",
                top_left_corner.as_raw(),
                top_left_corner.width() as u32,
                top_left_corner.height() as u32,
                image::ColorType::Rgba8,
            ).unwrap();
        }
    }
}

pub fn storage_mut(&mut self) -> Option<&mut (dyn Storage + 'static)>

A place where you can store custom data in a way that persists when you restart the app.

pub fn gl(&self) -> Option<&Arc<Context>>

A reference to the underlying glow (OpenGL) context.

This can be used, for instance, to:

Render things to offscreen buffers.

Read the pixel buffer from the previous frame (glow::Context::read_pixels).

Render things behind the egui windows.

Note that all egui painting is deferred to after the call to App::update (egui only collects egui::Shapes and then eframe paints them all in one go later on).

To get a glow context you need to compile with the glow feature flag, and run eframe using Renderer::Glow.

pub fn close(&mut self)

Tell eframe to close the desktop window.

The window will not close immediately, but at the end of the this frame.

Calling this will likely result in the app quitting, unless you have more code after the call to crate::run_native.

pub fn set_minimized(&mut self, minimized: bool)

Minimize or unminimize window. (native only)

pub fn focus(&mut self)

Bring the window into focus (native only). Has no effect on Wayland, or if the window is minimized or invisible.

This method puts the window on top of other applications and takes input focus away from them, which, if unexpected, will disturb the user.

pub fn request_user_attention(&mut self, kind: UserAttentionType)

If the window is unfocused, attract the user’s attention (native only).

Typically, this means that the window will flash on the taskbar, or bounce, until it is interacted with.

When the window comes into focus, or if None is passed, the attention request will be automatically reset.

See winit’s documentation for platform-specific effect details.

pub fn set_maximized(&mut self, maximized: bool)

Maximize or unmaximize window. (native only)

pub fn quit(&mut self)

👎Deprecated: Renamed close

Tell eframe to close the desktop window.

pub fn set_window_size(&mut self, size: Vec2)

Set the desired inner size of the window (in egui points).

pub fn set_window_title(&mut self, title: &str)

Set the desired title of the window.

pub fn set_decorations(&mut self, decorated: bool)

Set whether to show window decorations (i.e. a frame around you app).

If false it will be difficult to move and resize the app.

pub fn set_fullscreen(&mut self, fullscreen: bool)

Turn borderless fullscreen on/off (native only).

pub fn set_window_pos(&mut self, pos: Pos2)

set the position of the outer window.

pub fn drag_window(&mut self)

When called, the native window will follow the movement of the cursor while the primary mouse button is down.

Does not work on the web.

pub fn set_visible(&mut self, visible: bool)

Set the visibility of the window.

pub fn set_always_on_top(&mut self, always_on_top: bool)

On desktop: Set the window always on top.

(Wayland desktop currently not supported)

pub fn set_centered(&mut self)

On desktop: Set the window to be centered.

(Wayland desktop currently not supported)