Struct eframe::NativeOptions

pub struct NativeOptions {

    pub always_on_top: bool,
    pub maximized: bool,
    pub decorated: bool,
    pub drag_and_drop_support: bool,
    pub icon_data: Option<IconData>,
    pub initial_window_pos: Option<Pos2>,
    pub initial_window_size: Option<Vec2>,
    pub min_window_size: Option<Vec2>,
    pub max_window_size: Option<Vec2>,
    pub resizable: bool,
    pub transparent: bool,
    pub vsync: bool,
    pub multisampling: u16,
    pub depth_buffer: u8,
    pub stencil_buffer: u8,

}

Options controlling the behavior of a native window.

Only a single native window is currently supported.

Fields
always_on_top: bool

Sets whether or not the window will always be on top of other windows.

maximized: bool

Show window in maximized mode

decorated: bool

On desktop: add window decorations (i.e. a frame around your app)? If false it will be difficult to move and resize the app.

drag_and_drop_support: bool

On Windows: enable drag and drop support. Drag and drop can not be disabled on other platforms.

See winit’s documentation for information on why you might want to disable this on windows.

icon_data: Option<IconData>

The application icon, e.g. in the Windows task bar etc.

initial_window_pos: Option<Pos2>

The initial (inner) position of the native window in points (logical pixels).

initial_window_size: Option<Vec2>

The initial inner size of the native window in points (logical pixels).

min_window_size: Option<Vec2>

The minimum inner window size

max_window_size: Option<Vec2>

The maximum inner window size

resizable: bool

Should the app window be resizable?

transparent: bool

On desktop: make the window transparent. You control the transparency with App::clear_color(). You should avoid having a egui::CentralPanel, or make sure its frame is also transparent.

vsync: bool

Turn on vertical syncing, limiting the FPS to the display refresh rate.

The default is true.

multisampling: u16

Set the level of the multisampling anti-aliasing (MSAA).

Must be a power-of-two. Higher = more smooth 3D.

A value of 0 turns it off (default).

egui already performs anti-aliasing via “feathering” (controlled by egui::epaint::TessellationOptions), but if you are embedding 3D in egui you may want to turn on multisampling.

depth_buffer: u8

Sets the number of bits in the depth buffer.

egui doesn’t need the depth buffer, so the default value is 0.

stencil_buffer: u8

Sets the number of bits in the stencil buffer.

egui doesn’t need the stencil buffer, so the default value is 0.