Crate eframe

eframe - the egui framework crate

If you are planning to write an app for web or native, and want to use egui for everything, then eframe is for you!

To get started, see the examples. To learn how to set up eframe for web and native, go to https://github.com/emilk/eframe_template/ and follow the instructions there!

In short, you implement App (especially App::update) and then call crate::run_native from your main.rs, and/or use eframe::WebRunner from your lib.rs.

Usage, native:

use eframe::egui;

fn main() {
    let native_options = eframe::NativeOptions::default();
    eframe::run_native("My egui App", native_options, Box::new(|cc| Box::new(MyEguiApp::new(cc))));
}

#[derive(Default)]
struct MyEguiApp {}

impl MyEguiApp {
    fn new(cc: &eframe::CreationContext<'_>) -> Self {
        // Customize egui here with cc.egui_ctx.set_fonts and cc.egui_ctx.set_visuals.
        // Restore app state using cc.storage (requires the "persistence" feature).
        // Use the cc.gl (a glow::Context) to create graphics shaders and buffers that you can use
        // for e.g. egui::PaintCallback.
        Self::default()
    }
}

impl eframe::App for MyEguiApp {
   fn update(&mut self, ctx: &egui::Context, frame: &mut eframe::Frame) {
       egui::CentralPanel::default().show(ctx, |ui| {
           ui.heading("Hello World!");
       });
   }
}

Usage, web:

use wasm_bindgen::prelude::*;

/// Your handle to the web app from JavaScript.
#[derive(Clone)]
#[wasm_bindgen]
pub struct WebHandle {
    runner: eframe::WebRunner,
}

#[wasm_bindgen]
impl WebHandle {
    /// Installs a panic hook, then returns.
    #[allow(clippy::new_without_default)]
    #[wasm_bindgen(constructor)]
    pub fn new() -> Self {
        // Redirect [`log`] message to `console.log` and friends:
        eframe::WebLogger::init(log::LevelFilter::Debug).ok();

        Self {
            runner: eframe::WebRunner::new(),
        }
    }

    /// Call this once from JavaScript to start your app.
    #[wasm_bindgen]
    pub async fn start(&self, canvas_id: &str) -> Result<(), wasm_bindgen::JsValue> {
        self.runner
            .start(
                canvas_id,
                eframe::WebOptions::default(),
                Box::new(|cc| Box::new(MyEguiApp::new(cc))),
            )
            .await
    }

    // The following are optional:

    /// Shut down eframe and clean up resources.
    #[wasm_bindgen]
    pub fn destroy(&self) {
        self.runner.destroy();
    }

    /// Example on how to call into your app from JavaScript.
    #[wasm_bindgen]
    pub fn example(&self) {
        if let Some(app) = self.runner.app_mut::<MyEguiApp>() {
            app.example();
        }
    }

    /// The JavaScript can check whether or not your app has crashed:
    #[wasm_bindgen]
    pub fn has_panicked(&self) -> bool {
        self.runner.has_panicked()
    }

    #[wasm_bindgen]
    pub fn panic_message(&self) -> Option<String> {
        self.runner.panic_summary().map(|s| s.message())
    }

    #[wasm_bindgen]
    pub fn panic_callstack(&self) -> Option<String> {
        self.runner.panic_summary().map(|s| s.callstack())
    }
}

Simplified usage

If your app is only for native, and you don’t need advanced features like state persistence, then you can use the simpler function run_simple_native.

Feature flags

Re-exports

pub use egui;

pub use egui::emath;

pub use egui::epaint;

pub use egui_glow;

pub use glow;

Structs

CreationContext

Data that is passed to AppCreator that can be used to setup and initialize your app.

EventLoopBuilder

Object that allows building the event loop.

Frame

Represents the surroundings of your app.

IconData

Image data for an application icon.

IntegrationInfo

Information about the integration passed to the use app each frame.

NativeOptions

Options controlling the behavior of a native window.

WindowBuilder

Object that allows building windows.

WindowInfo

Information about the application’s main window, if available.

Enums

Error

The different problems that can occur when trying to run eframe.

HardwareAcceleration

Selects the level of hardware graphics acceleration.

Renderer

What rendering backend to use.

Theme

Dark or Light theme.

UserEvent

The custom even eframe uses with the winit event loop.

WebGlContextOption

WebGL Context options

Constants

APP_KEY

Storage key used for app

Traits

App

Implement this trait to write apps that can be compiled for both web/wasm and desktop/native using eframe.

Storage

A place where you can store custom data in a way that persists when you restart the app.

Functions

run_native

This is how you start a native (desktop) app.

run_simple_native

The simplest way to get started when writing a native app.

Type Aliases

AppCreator

This is how your app is created.

EventLoopBuilderHook

Hook into the building of an event loop before it is run

Result

Short for Result<T, eframe::Error>.

WindowBuilderHook

Hook into the building of a the native window.