Struct eframe::Frame

pub struct Frame { /* private fields */ }

Represents the surroundings of your app.

It provides methods to inspect the surroundings (are we on the web?), allocate textures, and change settings (e.g. window size).

Implementations

impl Frame

pub fn is_web(&self) -> bool

True if you are in a web environment.

pub fn info(&self) -> IntegrationInfo

Information about the integration.

pub fn storage(&self) -> Option<&dyn Storage>

A place where you can store custom data in a way that persists when you restart the app.

pub fn storage_mut(&mut self) -> Option<&mut (dyn Storage + 'static)>

A place where you can store custom data in a way that persists when you restart the app.

pub fn gl(&self) -> &Rc<Context>

A reference to the underlying glow (OpenGL) context.

This can be used, for instance, to:

Render things to offscreen buffers.

Read the pixel buffer from the previous frame (glow::Context::read_pixels).

Render things behind the egui windows.

Note that all egui painting is deferred to after the call to App::update (egui only collects egui::Shapes and then eframe paints them all in one go later on).

pub fn quit(&mut self)

Signal the app to stop/exit/quit the app (only works for native apps, not web apps). The framework will not quit immediately, but at the end of the this frame.

pub fn set_window_size(&mut self, size: Vec2)

Set the desired inner size of the window (in egui points).

pub fn set_window_title(&mut self, title: &str)

Set the desired title of the window.

pub fn set_decorations(&mut self, decorated: bool)

Set whether to show window decorations (i.e. a frame around you app). If false it will be difficult to move and resize the app.

pub fn set_window_pos(&mut self, pos: Pos2)

set the position of the outer window

pub fn drag_window(&mut self)

When called, the native window will follow the movement of the cursor while the primary mouse button is down.

Does not work on the web.